implementa tratamento de erros

+++ src/main/java/com/devsuperior/gscommerce/service/ProductService.java
@@ -1,15 +1,19 @@
 package com.devsuperior.gscommerce.service;
 
 
-import com.devsuperior.gscommerce.entities.dto.ProductRequestDTO;
+import com.devsuperior.gscommerce.dto.ProductRequestDTO;
 import com.devsuperior.gscommerce.entities.Product;
-import com.devsuperior.gscommerce.entities.dto.ProductResponseDTO;
-import com.devsuperior.gscommerce.exceptions.ResourceNotFoundException;
+import com.devsuperior.gscommerce.dto.ProductResponseDTO;
 import com.devsuperior.gscommerce.mapper.ProductMapper;
 import com.devsuperior.gscommerce.repositories.ProductRepository;
+import com.devsuperior.gscommerce.service.exceptions.CustomError;
+import com.devsuperior.gscommerce.service.exceptions.DataBaseException;
+import com.devsuperior.gscommerce.service.exceptions.ResourceNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -54,14 +58,21 @@ public class ProductService {
     @Transactional(readOnly = true)
     public ProductResponseDTO findById(Long id){
         Product entity = productRepository.findById(id)
-                .orElseThrow(()-> new ResourceNotFoundException("Id nao encontrado "+ id));
+                .orElseThrow(()-> new ResourceNotFoundException("Recurso nao encontrado "+ id));
 
         return ProductMapper.toDTO(entity);
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DataBaseException("Falha de integridade referencial");
+        }
     }
